Uploaded image of a sample bracelet. Uploaded Uploaded Uploaded image of a sample ring. Created a new page for the cart to display items  in the cart.
Updated the layout to include a Navbar
component.
Updated the main page to include a welcome message
 and a link to the products page.
Created a new page to display products.
Created a new page for the wishlist to display
 items that users might want to purchase later.
Created a Navbar component to navigate between
 different pages of the application.
Created a ProductCard component to display
 individual products with their details.
Created a dummy data file to simulate product data
	for development purposes.
Created a utility file for cart-related functions
to manage items in the cart.
Created a utility file for product-related functions
to manage product data.
Created a Tailwind CSS configuration file to
style the application using Tailwind CSS.
Updated the carts utility file to include
type definitions for cart items.